+++ frontend/src/pages/SuburbReportPage/SuburbReportPage.tsx
@@ -1,5 +1,6 @@
 import { Box, styled } from '@mui/material';
-import LifestyleAccessibilitySection from './components/LifestyleAccessibilitySection';
+import AccountBalanceIcon from '@mui/icons-material/AccountBalance';
+import MetricCardsSection from './components/MetricCardsSection';
 
 const PageContainer = styled(Box)(({ theme }) => ({
   maxWidth: '1440px',
@@ -7,13 +8,66 @@ const PageContainer = styled(Box)(({ theme }) => ({
   justifyContent: 'center',
   margin: '0 auto',
   padding: theme.spacing(8),
-  border: '1px solid black',
 }));
+
+const metricCardsData = [
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+  {
+    icon: <AccountBalanceIcon />,
+    title: 'Lifestyle Accessibility',
+    value: '80%',
+    subtitle: '12 months',
+  },
+];
 
 const SuburbReportPage = () => {
   return (
     <PageContainer>
-      <LifestyleAccessibilitySection />
+      <MetricCardsSection
+        title="Lifestyle Accessibility"
+        data={metricCardsData}
+      />
     </PageContainer>
   );
 };
