--- frontend/src/pages/SuburbReportPage/components/LifestyleAccessibilitySection/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LifestyleAccessibilitySection';
