+++ frontend/src/pages/SuburbReportPage/components/MetricCardsSection/components/MetricCard/index.ts
