+++ frontend/src/pages/SuburbReportPage/components/MetricCardsSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MetricCardsSection';
